LayoutTests/imported/w3c/web-platform-tests/custom-elements/Document-createElement-customized-builtins-expected.txt


PASS document.createElement must create an instance of autonomous custom elements when it has is attribute
FAIL document.createElement()'s second argument is to be ignored when it's a string assert_true: expected true got false
FAIL document.createElement must report an exception thrown by a custom built-in element constructor assert_throws_exactly: function "function () { new ThrowCustomBuiltinElement; }" threw object "TypeError: Illegal constructor" but we expected it to throw object "[object Object]"
FAIL document.createElement with unknown "is" value should create "undefined" state element assert_true: Undefined element is upgraded on connecting to a document expected true got false
